+++ EntityLib/EntityLibPy/EntityLibPy.cpp
@@ -43,7 +43,7 @@ namespace pybind11
 namespace py = pybind11;
 using namespace Ent;
 
-using Value = mapbox::util::variant<Null, std::string, float, int64_t, bool, EntityRef>;
+using Value = mapbox::util::variant<Null, std::string, double, int64_t, bool, EntityRef>;
 
 Value getValue(Ent::Node& node)
 {
@@ -83,11 +83,11 @@ struct Converter<IO, IO>
 };
 
 template <>
-struct Converter<int64_t, float>
+struct Converter<int64_t, double>
 {
-    float operator()(int64_t const& val) const
+    double operator()(int64_t const& val) const
     {
-        return static_cast<float>(val);
+        return static_cast<double>(val);
     }
 };
 
@@ -125,7 +125,7 @@ void setValue(Ent::Node& node, Value const& val)
         node.setInt(mapbox::util::apply_visitor(GetValue<int64_t>{}, val));
         break;
     case Ent::DataType::number:
-        node.setFloat(mapbox::util::apply_visitor(GetValue<float>{}, val));
+        node.setFloat(mapbox::util::apply_visitor(GetValue<double>{}, val));
         break;
     case Ent::DataType::string:
         node.setString(mapbox::util::apply_visitor(GetValue<std::string>{}, val).c_str());
@@ -342,7 +342,7 @@ PYBIND11_MODULE(EntityLibPy, ent)
         .def("is_default", [](Node* node) { return node->isDefault(); })
         .def("get_type_name", [](Node* node) { return node->getTypeName(); })
         .def_property("value", getValue, setValue)
-        .def("set_float", [](Node* node, float val) { return node->setFloat(val); })
+        .def("set_float", [](Node* node, double val) { return node->setFloat(val); })
         .def("set_int", [](Node* node, int64_t val) { return node->setInt(val); })
         .def("set_string", [](Node* node, char const* str) { return node->setString(str); })
         .def("set_bool", [](Node* node, bool val) { return node->setBool(val); })
